\name{countMatSC}
\alias{countMatSC}
\docType{data}
\title{
  The cell x gene matrix from single-cell RNA-seq. 
}
\description{
  A scRNA-seq dataset generated from placenta that has been analyzed with CellRanger and used to develop EmptyDrops. We have subsampled the genes from the real datasets to generated the matrix. 
}
\usage{data("countMatSC")}
\format{
  A data frame with 1090 cells and 10000 genes. This is the subset of data obtained from single-cell RNAseq for package testing.  Each row represents one cell detected in single-cell RNA-seq, each column is one gene in detected cells. The values in the matrix are the raw read counts from single-cell RNAseq. 
  \describe{
    A data frame with 1090 cells and 10000 genes.This is the subset of data obtained from single-cell RNAseq for package testing.
  }
}
\details{
  This data frame is the generated by single-cell RNA-seq.
}

\examples{
  data(countMatSC)
}
\keyword{datasets}

