\name{negGene}
\alias{negGene}
\docType{data}
\title{
  Input negative file. 
  }
\description{
  A data frame with 510 different well IDs in which the cells treated with non-specific siRNAs.If users did not have the build-in negative controls, the non-expressed genes should be provided here.
  }
\usage{data("negGene")}
\format{
  A data frame with 510 different well IDs in which the cells treated with non-specific siRNAs.
  \describe{
    A data frame with 510 different well IDs in which the cells treated with non-specific siRNAs. These wells were served as negative control.
  }
}
\details{
  These wells were designed by the authors in the large-scale screen.
}

\examples{
  data(negGene)
}
\keyword{datasets}
