\name{nonExpGene}
\alias{nonExpGene}
\docType{data}
\title{
  Input internal negative control file. 
  }
\description{
  A data frame with 722 different well IDs in which the cells treated with siRNAs targeting to non-expressed genes in HeLa cells.It the subset of total non-expressed 
  genes in HeLa cells.
  }
\usage{data("nonExpGene")}
\format{
  A data frame with 722 different well IDs in which the cells treated with siRNAs targeting to non-expressed genes in HeLa cells.
  \describe{
    A data frame with 722 different well IDs in which the cells treated with siRNAs targeting to non-expressed genes in HeLa cells. These wells were served as internal negative controls.
  }
}
\details{
  These non-expressed genes can be obtained from a prior expression profile.
}

\examples{
  data(nonExpGene)
}
\keyword{datasets}
