\name{posGene}
\alias{posGene}
\docType{data}
\title{
  Input positive file. 
  }
\description{
  A data frame with 299 different well IDs in which the cells treated with siRNAs targeting to PTB.If users didn't have the build-in positive controls, choose the parameters -withoutsvm and the filename can use any name such as 'NA'.
  }
\usage{data("negGene")}
\format{
  A data frame with 299 different well IDs in which the cells treated with siRNAs targeting to PTB.
  \describe{
    A data frame with 299 different well IDs in which the cells treated with siRNAs targeting to PTB. These wells were served as positive control.
  }
}
\details{
  These wells were designed by the authors in the large-scale screen.
}

\examples{
  data(posGene)
}
\keyword{datasets}
