\name{SVMcurve}
\alias{SVMcurve}
\docType{data}
\title{
  The SVM curve lines in Zeta-plot. 
}
\description{
  The SVM curves were calculated from raw input matrix files. They were designed to maximally seperate the positive and negative genes. 
  }
\usage{data("SVMcurve")}
\format{
  A data frame with 24 rows and 4 features.
  \describe{
    A data frame with 24 rows and 4 features.The first column is the bins cut-offs for decresed direction. The second column is the values of percentage with different cut-offs in column 1.
    The third column is the bins cut-offs for increased direction. The fourth column is the values of percentage with different cut-offs in column 3.
  }
}
\details{
  This data frame is the generated by SVM.R.
}

\examples{
  data(SVMcurve)
}
\keyword{datasets}

