\name{ZseqList}
\alias{ZseqList}
\docType{data}
\title{
  The bin size for Zeta calculation. 
  }
\description{
  A data frame with 11 different cut-offs and 2 directions. We divided the ranges of input values into bins. The number of bins is determined by the users.

  }
\usage{data("ZseqList")}
\format{
  A data frame with 11 different cut-offs and 2 directions.
  \describe{
    A data frame with 11 different cut-offs and 2 directions.We divided the ranges of input values into bins. The number of bins is determined by the users.
  }
}
\details{
  This data frame is the generated by EventCoverage.R.
}

\examples{
  data(ZseqList)
}
\keyword{datasets}
